refactor: move abbreviateName into lib/utils
Closes #21
